Organize data into lib/data
