+++ src/components/Controls/AddSmallBodyModal.tsx
@@ -17,7 +17,7 @@ import { IconChevronDown, IconX } from '@tabler/icons-react';
 import { UseQueryResult } from '@tanstack/react-query';
 import { useEffect, useMemo } from 'react';
 import { useSmallBodies } from '../../hooks/queries/useSmallBodies.ts';
-import { ASTEROIDS } from '../../lib/bodies.ts';
+import { ASTEROIDS } from '../../lib/data/bodies.ts';
 import { asCelestialBodyId, CelestialBody, CelestialBodyId, CelestialBodyType } from '../../lib/types.ts';
 import { nameToId, notNullish } from '../../lib/utils.ts';
