+++ src/components/Controls/SelectOmnibox.tsx
@@ -4,8 +4,9 @@ import { IconSearch } from '@tabler/icons-react';
 import { useMemo, useState } from 'react';
 import { useModifierKey } from '../../hooks/useModifierKey.ts';
 import { LABEL_FONT_FAMILY } from '../../lib/canvas.ts';
-import { ORBITAL_REGIMES, orbitalRegimeDisplayName } from '../../lib/regimes.ts';
-import { SPACECRAFT, SPACECRAFT_ORGANIZATIONS } from '../../lib/spacecraft.ts';
+import { SPACECRAFT_ORGANIZATIONS } from '../../lib/data/organizations.ts';
+import { ORBITAL_REGIMES, orbitalRegimeDisplayName } from '../../lib/data/regimes.ts';
+import { SPACECRAFT } from '../../lib/data/spacecraft.ts';
 import { Settings, UpdateSettings } from '../../lib/state.ts';
 import { CelestialBody } from '../../lib/types.ts';
 import { celestialBodyTypeDescription } from '../../lib/utils.ts';
