+++ src/components/FactSheet/CelestialBodyFactSheet.tsx
@@ -3,9 +3,9 @@ import { memo, ReactNode } from 'react';
 import { useFactsStream } from '../../hooks/queries/useFactsStream.ts';
 import { useDisplaySize } from '../../hooks/useDisplaySize.ts';
 import { useFactSheetPadding } from '../../hooks/useFactSheetPadding.ts';
-import { g } from '../../lib/bodies.ts';
+import { g } from '../../lib/data/bodies.ts';
+import { SPACECRAFT_BY_BODY_ID } from '../../lib/data/spacecraft.ts';
 import { orbitalPeriod, surfaceGravity } from '../../lib/physics.ts';
-import { SPACECRAFT_BY_BODY_ID } from '../../lib/spacecraft.ts';
 import { UpdateSettings } from '../../lib/state.ts';
 import { CelestialBody, CelestialBodyType } from '../../lib/types.ts';
 import { celestialBodyTypeDescription, humanDistanceUnits, humanTimeUnits, pluralize } from '../../lib/utils.ts';
