+++ src/components/FactSheet/Organization/OtherSpacecraftOrganizations.tsx
@@ -1,6 +1,6 @@
 import { Group, Stack, Title } from '@mantine/core';
 import { useFactSheetPadding } from '../../../hooks/useFactSheetPadding.ts';
-import { SPACECRAFT_ORGANIZATIONS } from '../../../lib/spacecraft.ts';
+import { SPACECRAFT_ORGANIZATIONS } from '../../../lib/data/organizations.ts';
 import { UpdateSettings } from '../../../lib/state.ts';
 import { SpacecraftOrganization } from '../../../lib/types.ts';
 import { SpacecraftOrganizationPill } from '../Spacecraft/SpacecraftOrganizationPill.tsx';
